So this fixes a couple of things...
First, there was an issue where after we saved, the document thought that the files on disk were newer, so it would re-read the document. This caused issues, because the re-read, on updating the document's storage, would interrupt certain editing operations, which is not what we wanted.

The first culprit for this was that I wasn't quite doing file wrappers correctly. I wasn't holding onto and modifying the wrapper originally read from disk. This seemed like a likely culprit, but that didn't fix the problem. Still, I'm leaving that change in, because it seems more correct.

I then noticed that when removing and re-adding a child wrapper from the top level document wrapper that the modification date on the held onto wrapper was cleared. It seems like the root bug is that when AppKit is done writing, it should be updating the wrappers to match the modification dates that're on disk, but that doesn't seem to be happening. To work around this, I wrote a convenience method that compares to file wrappers and their contents and if they have the same content with the same modification dates, then we say they're the same. When they're the same, we skip re-reading from disk, and just take ownership of the new wrapper.

There may be one more bug, but it's much harder to reproduce, so I'm unsure if it's still happening. Basically, that bug has to do with the document thinking the file has been modified on disk and is newer than the version that's trying to be saved. I think this is related to the above, but I'm not sure. I'm also not sure how the above changes will affect this additional bug. Anyways, if that bug comes up again, I'll investigate and fix separately.